requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pydantic~=1.8.2
numpy~=1.21.3
pytest~=6.2.5
fluentcheck~=0.2.0
pandas~=1.3.4
requests~=2.26.0
wfdb~=3.4.1